Fixes invalid `module` field in package.json
`module` referred to `src/index.js`, which does not exist. Point to `lib/index.js` instead.

Without fixing this, tools like rollup would give an error when trying to include as part of
a bundle. Alternatively, could probably just remove the `module` field entirely (since it is
the same as `main` in this case)